refactor(blobstorage): split settings page into feature components
Move the 800-line settings form out of the page file into
web/src/features/blobstorage-integration/components/: a form shell plus
field-group components (storage provider, export schedule, export
source, field groups, gzip toggle) and a status section. Replace the
duplicated defaults + manual reset useEffect with a shared
buildBlobStorageFormValues() driven through react-hook-form's reactive
`values` option (keepDirtyValues preserved).
